utils/utils-requirements.txt
git+http://github.com/Datera/python-sdk
openstacksdk
